Use ES6 + updates
Also adds the ability to customize which attribute is read.
If we found an attribute, we bail out. This is compatible with the old
version since we only checked 'value' and 'content' which should be on
empty elements.

To allow this change, we also needed to change the internal data
structure of selector: users can now use an object representation to
customize the attribute.